extraction per code review
